+++ frontend/src/components/feedbacks/Feedback.tsx
@@ -3,7 +3,7 @@ import { FeedbackType } from 'types';
 
 import useFeedback from 'hooks/useFeedback';
 
-import Button from './@commons/Button';
+import Button from '../@commons/Button';
 
 const Feedback = ({ userFeedback, requestFeedbackLookup }: FeedbackProps) => {
   const { feedbackDelete } = useFeedback();
