+++ frontend/src/components/feedbacks/LevellogReport.tsx
@@ -1,4 +1,4 @@
-import { FeedbackAddContainer } from './@commons/Style';
+import { FeedbackAddContainer } from '../@commons/Style';
 
 const LevellogReport = () => {
   return (
