+++ frontend/src/pages/FeedbackAdd.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import LevellogFeedback from 'components/LevellogFeedback';
-import LevellogReport from 'components/LevellogReport';
+import LevellogFeedback from 'components/feedbacks/LevellogFeedback';
+import LevellogReport from 'components/feedbacks/LevellogReport';
 
 const FeedbackAdd = () => {
   return (
